requirements.txt
Django>=4.2.9,<4.2.10
djangorestframework >= 3.14.0,<3.15
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.27.1,<0.28
Pillow>=10.2.0,<10.3.0
uwsgi>=2.0.23,<2.1
django-cors-headers>=4.3.1,<4.4
python-dotenv>=1.0.1, <1.1
langchain>=0.1.9, < 0.2
openai>=1.13.3, <1.14
